95% BusMapFragment finished (still needs states reset) and stripped Activity completely
